adding learn cards api
